// src/interfaces/ActivityInterfaces.tsx
import {Asset} from 'react-native-image-picker';
import {SimpleDiscipline} from './DisciplineInterfaces';
import {Address} from './LocationInterfaces';

export interface CreatedActivitiesResponse {
  activities: ActivityListItemData[];
}

export interface InterestedActivitiesResponse {
  liked: ActivityListItemData[];
  inscribed: ActivityListItemData[];
}

export interface ActivityDetails extends ActivityListItemData {
  phone: string;
  email: string;
  description: string;
  //instagram: string
}

export interface ActivityListItemData {
  id: number;
  name: string;
  price: number;
  //avRating: number;
  startDate: Date;
  endDate: Date;
  plan: Plan[];
  address: string;
  maxQuota: number;
  latitude: number;
  longitude: number;
  images: {url: string}[];
  createdBy: CreatedBy[];
  disciplines: Discipline[];
  enrolled: number;
  deleted?: boolean;
}

export interface Plan {
  startTime: Date;
  endTime: Date;
  days: Week[];
}

export enum Week {
  DOMINGO = 'DOMINGO',
  LUNES = 'LUNES',
  MARTES = 'MARTES',
  MIERCOLES = 'MIERCOLES',
  JUEVES = 'JUEVES',
  VIERNES = 'VIERNES',
  SABADO = 'SABADO',
}

// Generated by https://quicktype.io

export interface ActivityControlData {
  id?: number;
  name: string;
  description: string;
  price?: number;
  startDate: Date;
  endDate: Date;
  location: Address;
  maxQuota: number;
  disciplines: Discipline[];
  plan?: Plan[];
  images?: Image[];
  repeatable: boolean;
}

export interface Discipline {
  id: number;
  name: string;
}

export type Image = Asset | {id: number; url: string};

// Generated by https://quicktype.io

export interface CreatedBy {
  id: number;
  name: string;
  lastName: string;
  avRating: number | null;
  profilePic: {url: string} | null;
}

export interface SearchActivityRequestData {
  term?: string;
  minPrice?: number;
  maxPrice?: number;
  afterSetDays?: number;
  beforeSetDays?: number;
  weekDays?: string;
  // minRating?: number;
  repeatable?: boolean;
  maxQuota?: number;
  minQuota?: number;
  discipline?: SimpleDiscipline;
}

export interface ActivityMapMarkerData {
  id: number;
  name: string;
  address: string;
  maxQuota: number;
  enrolled: number;
  createdBy: CreatedBy[];
  startDate: Date;
  endDate: Date;
  plan: Plan[];
  disciplines: Discipline[];
  price: number;

  latitude: number;
  longitude: number;
}

//TODO: las actividades sin plan tmb suceden en días de la semana
//TODO: todo lo de los eventos!!!!!!

// Generated by https://quicktype.io